# swagger.yml
openapi: 3.0.3
info:
  title: Systeme de Pointage API
  description: Documentation des endpoints disponibles pour l'authentification, le pointage, le pilotage des cours et la gestion academique.
  version: 1.1.0
servers:
  - url: http://127.0.0.1:5000
    description: Serveur local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    kioskKeyAuth:
      type: apiKey
      in: header
      name: X-Attendance-Kiosk-Key
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          example: etudiant@test.com
        password:
          type: string
          example: Password123!
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        role:
          type: string
        user_id:
          type: integer
        must_change_password:
          type: boolean
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
        new_password:
          type: string
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          example: etudiant@test.com
    ResetPasswordRequest:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
          description: Token recu par email, extrait du lien de reinitialisation.
        new_password:
          type: string
          minLength: 8
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Authentification]
      summary: Connecter un utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion reussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Requete invalide
        '401':
          description: Identifiants invalides
        '403':
          description: Compte desactive
  /api/v1/auth/me:
    get:
      tags: [Authentification]
      summary: Recuperer le profil connecte
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil retourne
        '401':
          description: Non autorise
  /api/v1/auth/change-password:
    post:
      tags: [Authentification]
      summary: Changer le mot de passe apres premiere connexion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Mot de passe mis a jour
        '400':
          description: Requete invalide
        '401':
          description: Mot de passe actuel incorrect
  /api/v1/auth/forgot-password:
    post:
      tags: [Authentification]
      summary: Demander un lien de reinitialisation de mot de passe (etudiant, professeur, responsable, admin)
      description: >-
        Disponible pour les 4 roles (tous stockes dans la meme table Utilisateur).
        Reponse volontairement identique (200, message generique) que l'email
        corresponde ou non a un compte, et meme si le compte est desactive, pour ne
        pas permettre l'enumeration d'adresses email enregistrees. Si un compte actif
        existe, un email est envoye avec un lien vers FRONTEND_PASSWORD_RESET_URL
        contenant un token signe valable 30 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reponse generique (email envoye si le compte existe et est actif)
        '400':
          description: Email manquant
  /api/v1/auth/reset-password:
    post:
      tags: [Authentification]
      summary: Choisir un nouveau mot de passe a partir du token recu par email
      description: >-
        Le token est a usage unique : il embarque une empreinte du hash du mot de
        passe courant, donc toute reutilisation (ou tout changement de mot de passe
        entre-temps par un autre canal) invalide automatiquement les tokens deja
        emis, sans necessiter de table de revocation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Mot de passe reinitialise avec succes
        '400':
          description: token_invalide, token_expire, token_deja_utilise ou mot de passe trop court
        '403':
          description: Compte desactive
...